feat(generic parsers, normalizer): adds generic parsers for dates, strings + data normalizer
The data we get is arriving with some normalization issues: prop fields with space that prevents us
from referencing it directly, and null fields with text set to 'NULL'. This commit aims to normalize
the data to minimalize errors in code - as well as provide simple props already transformed for
final display (such as normalized.startDate).